This PostInit will be removed later.
Only added for debugging. But we can see the buffered action is not being set on the serverside which is why the animation plays, but the action does not actually occur. So we have to manipulate the player controller into picking up our buffering action when we press our attack key as if we had just performed a component action.